% Auto-generated documentation for function entries<- 
%  2021-06-02 11:12:18 
\name{entries<-}
\alias{entries<-}
\title{Update the Table Entries of a Plot-Ready or Plotted Table }
\description{
Update the table entries component of a plot-ready (\code{prTable}) or
plotted (\code{pltdTable}) table object.  This is an internal function,
not intended to be called by package users.
}
\usage{
entries(x) <- value
}
\arguments{
  \item{x}{A plot-ready (\code{prTable}) or plotted (\code{pltdTable}) table object.
  }
  \item{value}{A \code{prEntries} object containing the new plot-ready table entries.
  }
}
\value{
An object like \code{x}, with updated table entries.
}
\keyword{internal}
