% Auto-generated documentation for function hvrules<- 
%  2020-11-29 11:21:55 
\name{hvrules<-}
\alias{hvrules<-}
\title{Update the Hvrules of a Plot-Ready or Plotted Table }
\description{
Update the hvrules component of a plot-ready (\code{prTable}) or plotted
(\code{pltdTable}) table object.  This is an internal function, not
intended to be called by package users.
}
\usage{
hvrules(x) <- value
}
\arguments{
  \item{x}{A plot-ready (\code{prTable}) or plotted (\code{pltdTable}) table object.
  }
  \item{value}{A \code{prHvrules} object containing the new plot-ready hvrules.
  }
}
\value{
An object like \code{x}, with updated hvrules.
}
\keyword{internal}
