% Auto-generated documentation for function plot.pltdTable 
%  2021-06-02 11:12:19 
\name{plot.pltdTable}
\alias{plot.pltdTable}
\title{Display a \code{pltdtable} Object on a Graphics Device }
\description{
Display a \code{pltdTable} object on a graphics device.  This is an alias
for \code{print.pltdTable}.  This is an internal function, not intended to
be called by package users.
}
\usage{
\method{plot}{pltdTable}(x, ...)
}
\arguments{
  \item{x}{A \code{pltdTable} object, representing a table.
  }
  \item{...}{Additional arguments passed to \code{print.pltdTable}.
  }
}
\value{
\code{x}, invisibly.
}
\seealso{
\code{\link{print.pltdTable}}
}
\keyword{internal}
