% Auto-generated documentation for function print.summary.pltdTable 
%  2024-01-26 15:48:14.113213 
\name{print.summary.pltdTable}
\alias{print.summary.pltdTable}
\title{Print the Dimensions and Display Options of a Plotted Table }
\description{
Print method for \code{summary.pltdTable} objects.  The dimensions and
display options of a plotted table are printed.  This is an internal
function, not intended to be called directly by package users.
}
\usage{
\method{print}{summary.pltdTable}(x, ...)
}
\arguments{
  \item{x}{A \code{summary.pltdTable} object.
  }
  \item{...}{Additional arguments, ignored with a warning.  (Included for compatibility
with the generic.)
  }
}
\value{
\code{x} is returned, invisibly.
}
\keyword{internal}
