% Auto-generated documentation for function print.summary.textTable 
%  2020-11-29 11:21:55 
\name{print.summary.textTable}
\alias{print.summary.textTable}
\title{Print the Dimensions of a Table and Its Parts }
\description{
Print method for \code{summary.textTable} objects.  The dimensions of a
table and its parts are printed.  This is an internal function, not
intended to be called directly by package users.
}
\usage{
\method{print}{summary.textTable}(x, ...)
}
\arguments{
  \item{x}{A \code{summary.textTable} object.
  }
  \item{...}{Additional arguments, ignored with a warning.  (Included for compatibility
with the generic.)
  }
}
\value{
\code{x} is returned, invisibly.
}
\keyword{internal}
