% Auto-generated documentation for function prTable 
%  2021-06-02 11:12:19 
\name{prTable}
\alias{prTable}
\title{Create or Update a Fully Styled Table Ready for Plotting }
\description{
Create or update a \code{prTable} object, a fully styled (plot-ready)
table.  This is an S3 generic.  It and its methods are internal functions,
not intended to be called by package users.
}
\usage{
prTable(x, ...)
}
\arguments{
  \item{x}{An object to be converted to a plot-ready table.
  }
  \item{...}{Additional arguments passed to specific methods.
  }
}
\value{
An object of S3 class \code{prTable}.  See \code{prTable.prEntries} for
the structure of this object.
}
\keyword{internal}
